#72 refactor: 파일 구조 리팩터링
